Create NTriples class
